+++ src/apps/actionlint.rs
@@ -76,6 +76,10 @@ impl App for ActionLint {
             None => AnalyzeResult::NotIdentified,
         }
     }
+
+    fn allowed_versions(&self) -> Option<semver::VersionReq> {
+        None
+    }
 }
 
 fn cpu_text(cpu: Cpu) -> &'static str {
