+++ src/apps/dprint.rs
@@ -75,6 +75,10 @@ impl App for Dprint {
             None => AnalyzeResult::IdentifiedButUnknownVersion,
         }
     }
+
+    fn allowed_versions(&self) -> Option<semver::VersionReq> {
+        None
+    }
 }
 
 fn cpu_text(cpu: Cpu) -> &'static str {
