+++ src/apps/gh.rs
@@ -66,6 +66,10 @@ impl App for Gh {
             None => AnalyzeResult::IdentifiedButUnknownVersion,
         }
     }
+
+    fn allowed_versions(&self) -> Option<semver::VersionReq> {
+        None
+    }
 }
 
 fn cpu_text(cpu: Cpu) -> &'static str {
