+++ src/filesystem/mod.rs
@@ -1,4 +1,5 @@
 mod create_parent;
+mod find_file;
 mod find_global_install;
 mod has_extension;
 mod make_file_executable;
@@ -6,6 +7,7 @@ mod save_buffer;
 mod strip_filepath;
 
 pub use create_parent::create_parent;
+pub use find_file::find_file;
 pub use find_global_install::find_global_install;
 pub use has_extension::has_extension;
 pub use make_file_executable::make_file_executable;
